refactor: extract realtime file watcher into collection/watcher.rs (#577 item 17)
First step of splitting collection.rs's 5 bundled domains (#577): the
realtime notify-based watcher and its debounced batch-reconciliation
logic move into a `collection::watcher` submodule, re-exporting
`start_watcher`/`WatcherPauseGuard` from `collection.rs` to keep all
existing external call sites unchanged. Scan-time logic that's shared
between the watcher and the scanner (`reconcile_moved_songs`,
`unreachable_watched_roots`, `resolve_case_insensitive_path`) stays in
collection.rs for now.
